requirements.txt
python-telegram-bot==22.5
google-generativeai==0.3.2
aiosqlite==0.19.0
python-dotenv==1.0.0



